add listener file. Fix hinder bug.
Now track seems have a bug that left and right 's width is not the same.

+++ gizmo/src/main/java/com/ecnu/ooad/Constants.java
@@ -15,6 +15,6 @@ public class Constants {
     public static final int HINDER_WIDTH = 20;
     public static final int HINDER_HEIGHT = 2;
     public static final float RATE = 10f;
-    public static final int TRACK_WIDTH = 2;
+    public static final float TRACK_WIDTH = 2;
     public static final float TIME_STEP = 1f / 30f;
 }
